+++ lib/main.dart
@@ -1,11 +1,14 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
+import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:package_info_plus/package_info_plus.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:talker_flutter/talker_flutter.dart';
+import 'package:talker_riverpod_logger/talker_riverpod_logger.dart';
 
+import 'api/dio_client.dart';
 import 'app/app.dart';
-import 'core/theme/theme_provider.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -13,88 +16,27 @@ Future<void> main() async {
   final talker = TalkerFlutter.init();
   talker.debug('Talker started...');
 
+  await dotenv.load(fileName: ".env");
+
   final sharedPreferences = await SharedPreferences.getInstance();
+  final apiClient = YoutubeDioClient(talker);
 
   final config = Config(
       preferences: sharedPreferences,
-      talker: talker
+      talker: talker,
+      apiClient: apiClient
   );
 
-  runApp(Scope(config: config));
-}
+  DIContainer(config).register();
 
+  runApp(ProviderScope(
+    observers: [
+      TalkerRiverpodObserver(talker: talker),
+    ],
+    child: GuessSongApp(),
+  ));
 
-
-class MyHomePage extends ConsumerStatefulWidget {
-  const MyHomePage({super.key, required this.title});
-  final String title;
-
-  @override
-  ConsumerState<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends ConsumerState<MyHomePage> {
-  int _counter = 0;
-  String _version = '';
-
-  @override
-  Widget build(BuildContext context) {
-    final isDark = ref.watch(themeProvider).isDark;
-
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        title: Text(widget.title),
-      ),
-      body: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: <Widget>[
-          const Text('You have pushed the button this many times:'),
-          Text(
-            '$_counter',
-            style: Theme.of(context).textTheme.headlineMedium,
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              ElevatedButton(
-                  onPressed: _getVersion,
-                  child: Text('Get Version')
-              ),
-              SizedBox(width: 10),
-              Text('Version: $_version')
-            ],
-          ),
-          Container(
-            width: 300,
-            child: SwitchListTile(
-              title: const Text("Dark theme"),
-              value: isDark,
-              onChanged: (value) {
-                ref.read(themeProvider.notifier).setThemeBrightness(value);
-              },
-            ),
-          ),
-        ],
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _incrementCounter,
-        tooltip: 'Increment',
-        child: const Icon(Icons.add),
-      ),
-    );
-  }
-
-  void _incrementCounter() {
-    setState(() {
-      _counter++;
-    });
-  }
-
-  Future<void> _getVersion() async {
-    final info = await PackageInfo.fromPlatform();
-    setState(() {
-      _version = info.version;
-    });
-  }
+  FlutterError.onError = (details) {
+    talker.handle(details.exception, details.stack, 'Something went wrong..');
+  };
 }