+++ lib/repositories/youtube/models/youtube_video.dart
@@ -1,0 +1,13 @@
+class YoutubeVideo {
+  final String id;
+  final String title;
+  final String artist;
+  final String thumbnailUrl;
+
+  YoutubeVideo({
+    required this.id,
+    required this.title,
+    required this.artist,
+    required this.thumbnailUrl,
+  });
+}