+++ lib/repositories/youtube/youtube.dart
@@ -1,0 +1,2 @@
+export 'youtube_repository.dart';
+export 'youtube_repository_interface.dart';