+++ lib/repositories/youtube/youtube_repository_interface.dart
@@ -1,0 +1,5 @@
+import 'models/youtube_video.dart';
+
+abstract class IYoutubeRepository {
+  Future<List<YoutubeVideo>> searchSong(String query);
+}