+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.2.0] - 2025-09-09
+- Connected to YouTube API
+- Added Game Setup screen
+- Added Game screen
+- Added Game Summary screen
+- Added Settings screen
+
 ## [1.1.0] - 2025-09-05
 - Added icon and splash screen
 - Support for light and dark themes